// src/pages/DocumentsDetails/types.ts
export interface DocumentDetailsRow {
  documenttype: string;
  suppliernumber: string;
  documentserial: string;
  documentnumber: string;
  documentdate: string;
  suppliername: string;
  description: string;
  vehicle_nro: string;
  unit_measure_description: string;
  quantity: number;
  currency: string;
  unit_value: number;
  total_value: number;
  amount: number;
  taxamount: number;
  totalamount: number;
  driver: string;
  centercost: string;
}